Improve shopping cart perfomance
