+++ packages/evershop/src/modules/checkout/services/getCartByUUID.js
@@ -1,6 +1,4 @@
-const { select } = require('@evershop/postgres-query-builder');
-const { pool } = require('@evershop/evershop/src/lib/postgres/connection');
-const { Cart } = require('./cart/Cart');
+const { getCart } = require('./cart/Cart');
 
 module.exports = exports;
 
@@ -10,14 +8,6 @@ module.exports = exports;
  * @returns {Promise<Cart>}
  */
 exports.getCartByUUID = async (uuid) => {
-  const query = select().from('cart');
-  query.where('uuid', '=', uuid).and('status', '=', 1);
-  const data = await query.load(pool);
-  if (!data) {
-    return null;
-  } else {
-    const cart = new Cart({ ...data });
-    await cart.build();
-    return cart;
-  }
+  const cart = await getCart(uuid);
+  return cart;
 };
