+++ packages/evershop/src/modules/cod/bootstrap.js
@@ -1,20 +1,27 @@
-const { Cart } = require('../checkout/services/cart/Cart');
+const { addProcessor } = require('../../lib/util/registry');
 const { getSetting } = require('../setting/services/setting');
 
 module.exports = () => {
-  Cart.addField('payment_method', async function resolver(previousValue) {
-    const paymentMethod = this.dataSource?.payment_method ?? null;
-    if (paymentMethod !== 'cod') {
-      return previousValue;
-    } else {
-      // Validate the payment method
-      const codStatus = await getSetting('codPaymentStatus');
-      if (parseInt(codStatus, 10) !== 1) {
-        return previousValue;
-      } else {
-        delete this.errors.payment_method;
-        return paymentMethod;
-      }
-    }
+  addProcessor('cartFields', (fields) => {
+    fields.push({
+      key: 'payment_method',
+      resolvers: [
+        async function resolver(paymentMethod) {
+          if (paymentMethod !== 'cod') {
+            return paymentMethod;
+          } else {
+            // Validate the payment method
+            const codStatus = await getSetting('codPaymentStatus');
+            if (parseInt(codStatus, 10) !== 1) {
+              return null;
+            } else {
+              this.setError('payment_method', undefined);
+              return paymentMethod;
+            }
+          }
+        }
+      ]
+    });
+    return fields;
   });
 };
