+++ packages/evershop/src/modules/paypal/bootstrap.js
@@ -1,20 +1,28 @@
-const { Cart } = require('../checkout/services/cart/Cart');
+const { addProcessor } = require('../../lib/util/registry');
 const { getSetting } = require('../setting/services/setting');
 
 module.exports = () => {
-  Cart.addField('payment_method', async function resolver(previousValue) {
-    const paymentMethod = this.dataSource?.payment_method ?? null;
-    if (paymentMethod !== 'paypal') {
-      return previousValue;
-    } else {
-      // Validate the payment method
-      const paypalStatus = await getSetting('paypalPaymentStatus');
-      if (parseInt(paypalStatus, 10) !== 1) {
-        return previousValue;
-      } else {
-        delete this.errors.payment_method;
-        return paymentMethod;
-      }
-    }
+  addProcessor('cartFields', (fields) => {
+    fields.push({
+      key: 'payment_method',
+      resolvers: [
+        async function resolver(paymentMethod) {
+          // Do nothing if the payment method is not paypal
+          if (paymentMethod !== 'paypal') {
+            return paymentMethod;
+          } else {
+            // Validate the payment method
+            const paypalStatus = await getSetting('paypalPaymentStatus');
+            if (parseInt(paypalStatus, 10) !== 1) {
+              return null;
+            } else {
+              this.setError('payment_method', undefined);
+              return paymentMethod;
+            }
+          }
+        }
+      ]
+    });
+    return fields;
   });
 };
