+++ packages/evershop/src/modules/stripe/bootstrap.js
@@ -1,20 +1,27 @@
-const { Cart } = require('../checkout/services/cart/Cart');
+const { addProcessor } = require('../../lib/util/registry');
 const { getSetting } = require('../setting/services/setting');
 
 module.exports = () => {
-  Cart.addField('payment_method', async function resolver(previousValue) {
-    const paymentMethod = this.dataSource?.payment_method ?? null;
-    if (paymentMethod !== 'stripe') {
-      return previousValue;
-    } else {
-      // Validate the payment method
-      const stripeStatus = await getSetting('stripePaymentStatus');
-      if (parseInt(stripeStatus, 10) !== 1) {
-        return previousValue;
-      } else {
-        delete this.errors.payment_method;
-        return paymentMethod;
-      }
-    }
+  addProcessor('cartFields', (fields) => {
+    fields.push({
+      key: 'payment_method',
+      resolvers: [
+        async function resolver(paymentMethod) {
+          if (paymentMethod !== 'stripe') {
+            return paymentMethod;
+          } else {
+            // Validate the payment method
+            const stripeStatus = await getSetting('stripePaymentStatus');
+            if (parseInt(stripeStatus, 10) !== 1) {
+              return null;
+            } else {
+              this.setError('payment_method', undefined);
+              return paymentMethod;
+            }
+          }
+        }
+      ]
+    });
+    return fields;
   });
 };
